2018-1-6
目前主要完成了o2o项目前期的搭建和店铺注册功能的实现，主要遇到的问题是图片上传问题，由于源码是重构后的代码，而视频讲解的是重构前的代码，所以有些地方是不互通的，所以目前的代码是重构前能够完成图片上传的代码：
1.图片上传所需要的包都具备；
2.图片是通过流的形式上传的，解析的很完整；
3.可以看下店铺注册类中关于图片格式的相互转换过程好图片处理相关的类，对以后的图片处理有帮助

+++ o2o/src/main/java/com/akk/o2o/entity/Area.java
@@ -15,7 +15,7 @@ public class Area {
     /**
      * Id
      */
-    private Long areaId;
+    private Integer areaId;
 
     /**
      * 名称
